Upd logger test.
